警告を消す

+++ not/GameExplanation.cpp
@@ -34,7 +34,7 @@ void GameExplanation::NormalUpdate(const InputState& input)
 
 void GameExplanation::FadeOutUpdate(const InputState& input)
 {
-	fadeValue = 255 * (static_cast<int>(fadeTimer) / static_cast<int>(fadeIntarval));
+	fadeValue = static_cast<int>(255 * (static_cast<float>(fadeTimer) / static_cast<float>(fadeIntarval)));
 	if (++fadeTimer == fadeIntarval)
 	{
 		manager_.CangeScene(new KeyExplanationScene(manager_));
