+++ not/MixField.cpp
@@ -18,7 +18,7 @@ namespace
 
 void MixField::FadeInUpdate(const InputState& input)
 {
-	fadeValue = 255 * static_cast<int>(fadeTimer) / static_cast<int>(fadeInterval);
+	fadeValue = static_cast<int>(255 * static_cast<float>(fadeTimer) / static_cast<float>(fadeInterval));
 	if (--fadeTimer == 0)
 	{
 		updateFunc = &MixField::NormalUpdate;
@@ -385,7 +385,7 @@ void MixField::CountDownUpdate()
 
 void MixField::FadeOutUpdate(const InputState& input)
 {
-	fadeValue = 255 * static_cast<int>(fadeTimer) / static_cast<int>(fadeInterval);
+	fadeValue = static_cast<int>(255 * static_cast<float>(fadeTimer) / static_cast<float>(fadeInterval));
 	if (++fadeTimer == fadeInterval)
 	{
 		manager_.CangeScene(new MixGameOverScene(manager_));
