+++ not/PadField.cpp
@@ -18,7 +18,7 @@ namespace
 
 void PadField::FadeInUpdate(const InputState& input)
 {
-	fadeValue = 255 * static_cast<int>(fadeTimer) / static_cast<int>(fadeInterval);
+	fadeValue = static_cast<int>(255 * static_cast<float>(fadeTimer) / static_cast<float>(fadeInterval));
 	if (--fadeTimer == 0)
 	{
 		updateFunc = &PadField::NormalUpdate;
@@ -258,7 +258,7 @@ void PadField::CountDownUpdate()
 
 void PadField::FadeOutUpdate(const InputState& input)
 {
-	fadeValue = 255 * static_cast<int>(fadeTimer) / static_cast<int>(fadeInterval);
+	fadeValue = static_cast<int>(255 * static_cast<float>(fadeTimer) / static_cast<float>(fadeInterval));
 	if (++fadeTimer == fadeInterval)
 	{
 		manager_.CangeScene(new PadGameOverScene(manager_));
