+++ not/TitleScene.cpp
@@ -34,7 +34,7 @@ void TitleScene::NormalUpdate(const InputState& input)
 
 void TitleScene::FadeOutUpdate(const InputState& input)
 {
-	fadeValue = 255 * (static_cast<float>(fadeTimer) / static_cast<float>(fadeIntarval));
+	fadeValue = static_cast<int>(255 * (static_cast<float>(fadeTimer) / static_cast<float>(fadeIntarval)));
 	if (++fadeTimer == fadeIntarval)
 	{
 		manager_.CangeScene(new KeyExplanationScene(manager_));
